+++ app/src/main/java/com/sopt/now/data/request/RequestSignInDto.kt
@@ -1,0 +1,12 @@
+package com.sopt.now.data.request
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RequestSignInDto(
+    @SerialName("authenticationId")
+    val authenticationId: String,
+    @SerialName("password")
+    val password: String,
+)
