--- app/src/main/java/com/sopt/now/signIn/RequestSignInDto.kt
@@ -1,12 +1,0 @@
-package com.sopt.now.signIn
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class RequestSignInDto(
-    @SerialName("authenticationId")
-    val authenticationId: String,
-    @SerialName("password")
-    val password: String,
-)
