--- app/src/main/java/com/sopt/now/signIn/ResponseSignInDto.kt
@@ -1,12 +1,0 @@
-package com.sopt.now.signIn
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ResponseSignInDto(
-    @SerialName("code")
-    val status: Int,
-    @SerialName("message")
-    val message: String,
-)
