--- app/src/main/java/com/sopt/now/signIn/SignInState.kt
@@ -1,7 +1,0 @@
-package com.sopt.now.signIn
-
-data class SignInState(
-    val isSuccess: Boolean,
-    val message: String,
-    val memberId: String? = ""
-)
