--- app/src/main/java/com/sopt/now/signUp/SignUpState.kt
@@ -1,6 +1,0 @@
-package com.sopt.now.signUp
-
-data class SignUpState(
-    val isSuccess: Boolean,
-    val message: String
-)
