+++ app/src/main/java/com/sopt/now/ui/signIn/SignInState.kt
@@ -1,0 +1,7 @@
+package com.sopt.now.ui.signIn
+
+data class SignInState(
+    val isSuccess: Boolean,
+    val message: String,
+    val memberId: String? = ""
+)
