+++ app/src/main/java/com/sopt/now/ui/signUp/SignUpActivity.kt
@@ -1,0 +1,59 @@
+package com.sopt.now.ui.signUp
+
+import android.content.Intent
+import android.os.Bundle
+import android.widget.Toast
+import androidx.activity.viewModels
+import androidx.appcompat.app.AppCompatActivity
+import com.sopt.now.data.request.RequestSignUpDto
+import com.sopt.now.databinding.ActivitySignupBinding
+import com.sopt.now.ui.signIn.SignInActivity
+
+class SignUpActivity : AppCompatActivity() {
+
+    private val binding by lazy { ActivitySignupBinding.inflate(layoutInflater) }
+    private val viewModel by viewModels<SignUpViewModel>()
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(binding.root)
+        initViews()
+        initObserver()
+    }
+
+    private fun initViews() {
+        binding.btnSignUpSignUp.setOnClickListener {
+            viewModel.signUp(getSignUpRequestDto())
+        }
+    }
+
+    private fun initObserver() {
+        viewModel.liveData.observe(this) {
+            showToast(it.message)
+        }
+
+        viewModel.navigateToSignIn.observe(this) { navigate ->
+            if (navigate) {
+                val intent = Intent(this, SignInActivity::class.java)
+                startActivity(intent)
+                viewModel.doneNavigatingToSignIn()
+            }
+        }
+    }
+
+    private fun getSignUpRequestDto(): RequestSignUpDto {
+        val id = binding.etSignUpId.text.toString()
+        val password = binding.etSignUpPw.text.toString()
+        val nickname = binding.etSignUpName.text.toString()
+        val phoneNumber = binding.etSignUpPhoneNumber.text.toString()
+        return RequestSignUpDto(
+            authenticationId = id,
+            password = password,
+            nickname = nickname,
+            phone = phoneNumber
+        )
+    }
+
+    private fun showToast(message: String) {
+        Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
+    }
+}
