+++ app/src/main/java/com/sopt/now/ui/signUp/SignUpState.kt
@@ -1,0 +1,6 @@
+package com.sopt.now.ui.signUp
+
+data class SignUpState(
+    val isSuccess: Boolean,
+    val message: String
+)
